backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-multipart==0.0.6
httpx==0.25.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
pydantic==2.5.1
openai==1.3.3
python-dotenv==1.0.0
mysql-connector-python==8.2.0
PyJWT==2.8.0 